backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
lxml==5.2.0
numpy==1.26.4
scipy==1.13.0
httpx==0.27.0
pydantic==2.7.0
PyJWT==2.9.0
bcrypt==4.2.0

